+++ cmd/gogoodwe/semsapi/apilogin.go
@@ -2,7 +2,7 @@
 # Name: apiLogin - Logs in to the SEMS API
 # Author: Aaron Saikovski - asaikovski@outlook.com
 */
-package login
+package semsapi
 
 import (
 	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/authentication"
@@ -10,7 +10,7 @@ import (
 	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/utils"
 )
 
-// apiLogin -  Login to the API
+// SemsApiLogin -  Login to the GoodWe Sems Portal API
 func ApiLogin(UserLoginCreds *types.LoginCredentials, UserLoginResponse *types.LoginResponse) error {
 
 	// Do the login - update the pointer to the struct SemsResponseData
